--- gcn/functions/__init__.py
@@ -1,6 +1,0 @@
-from .GOF import GOF_Function, my_GOF_Function
-
-# def my_GOF_Function():
-# 	return my_GOF_Function()
-
-__all__ = ["GOF_Function","my_GOF_Function"]
