--- gcn/modules/__init__.py
@@ -1,3 +1,0 @@
-from .GConv import GConv, getGaborFilterBank
-
-__all__ = ['GConv', 'getGaborFilterBank']
